xconfig: Don't ignore interfaces without IPs
Some configurations don't actually need IP addresses, so interfaces
without IPs set shouldn't be excluded.  This commit adds a parameter
to getLinuxInterfaces() so that you can choose whether or not an
interface must have an IP to be of interest.  It currently defaults
to no-IP being fine.  In actuality, its callers should set this
based on context.  For example, I think a normal router doesn't
need it, but a dual stack host does.