Reduce CPU consumption / slow things down
There was a lot of fairly high frequency control traffic and such.
This turns a lot of it way down.  A lot of things now happen on
the order of every few seconds instead of on the order of several
times per second.

The new values weren't selected particularly scientifically.
Empirically, however, they make things much more reliable in
CPU-limited testing scenarios.